html/browsers/browsing-the-web/navigating-across-documents/008.html

<!doctype html>
<!--/
specs:
  - id: browsing-the-web.html#navigating-across-documents
    steps: [6, 11, 12]
  - id: browsing-the-web.html#javascript-protocol
  - id: form-control-infrastructure.html#concept-form-submit
  - id: form-control-infrastructure.html#submit-get-action
  - id: form-control-infrastructure.html#plan-to-navigate
  - id: form-control-infrastructure.html#planned-navigation

description: |
  1. This test creates a form with `target="test"` and an anchor with both an onclick and href.
  2. The test "clicks" the anchor, which will trigger the onclick operation, which will attempt to submit the form. The form's `action` is a `javascript:` url, which has special handling in step 12
  3. Navigating to the anchor's `href=""` will abort the `action`.

/-->
<title>Link with onclick form submit to javascript: url is aborted by href navigation</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe id="test" name="test"></iframe>
<form target="test" action="javascript:parent.postMessage('click', '*')"></form>
<a target="test" onclick="document.forms[0].submit()" href="href.html">Test</a>
<script>
var t = async_test();
t.step(function() {
  document.getElementsByTagName("a")[0].click();
});
window.form = document.getElementsByTagName("form")[0];
onmessage = t.step_func(function(e) {
  assert_equals(e.data, "href");
  t.done();
});
</script>
